style: Categorize files by class
Created 'myhomepage.dart' and 'sliderightroute.dart' files for better
readability and classification. And also deleted the unnecessary annotations.
